requirements.txt
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dateutil==2.8.2
selenium==4.15.2
flask==3.0.0 